transcription/requirements.txt
aiohttp==3.8.1
aiohttp-wsgi==0.10.0
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
deepgram-sdk==0.2.4
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
python-dotenv==0.19.2
websockets==10.2
Werkzeug==2.0.3
yarl==1.7.2